Replaced all float with CGFloat which fixes it working on 10.9.

+++ CTGradientView.h
@@ -13,7 +13,7 @@
 	{
 	CTGradient *myGradient;
 	
-	float angle;
+	CGFloat angle;
 	bool  isRadial;
 	}
